# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Hifive football api",
    "version": "1.0.0",
    "title": "HifiveFootball",
    "termsOfService": "http://api.hifivefootball.com/",
    "contact": {
      "email": "ddastudio@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.hifivefootball.com:5300",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/issue": {
      "post": {
        "tags": [
          "issue"
        ],
        "summary": "content",
        "description": "post content",
        "operationId": "postIssue",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "name": "boardid",
            "in": "formData",
            "description": "700 is issue board",
            "required": true,
            "type": "string",
            "default": 700
          },
          {
            "name": "title",
            "in": "formData",
            "description": "offset of",
            "required": true,
            "type": "string"
          },
          {
            "name": "content",
            "in": "formData",
            "description": "offset of",
            "required": false,
            "type": "string"
          },
          {
            "name": "ip",
            "in": "formData",
            "description": "offset of",
            "required": false,
            "type": "string"
          },
          {
            "name": "tags",
            "in": "formData",
            "description": "json [\"ronaldo\", \"messi\"]",
            "required": false,
            "type": "string"
          },
          {
            "name": "imgs",
            "in": "formData",
            "description": "json",
            "required": false,
            "type": "string"
          },
          {
            "name": "bodytype",
            "in": "formData",
            "description": "1 html , 2 text , 3 external link",
            "required": false,
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "default": 1
          },
          {
            "name": "platform",
            "in": "formData",
            "description": "1 android 2 ios 3 web, 아마 안쓸것임",
            "required": false,
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "default": 3
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContentHeader"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/issue/{id}": {
      "get": {
        "tags": [
          "issue"
        ],
        "summary": "Finds issue by id",
        "description": "Multiple status values can be provided with comma separated strings",
        "operationId": "getIssue",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "description": "issueid",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Issue"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/ApiResponse2"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ContentHeader": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "arena_id": {
          "type": "integer"
        },
        "deleted": {
          "type": "integer"
        },
        "reported": {
          "type": "integer"
        },
        "comments": {
          "type": "integer"
        },
        "scraps": {
          "type": "integer"
        },
        "board_id": {
          "type": "integer"
        },
        "preview": {
          "type": "string"
        },
        "likers": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        },
        "bodytype": {
          "type": "integer"
        },
        "celltype": {
          "type": "integer"
        },        
        "link": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      },
      "xml": {
        "name": "ContentHeader"
      }
    },
    "Issue": {
      "type": "object",
      "properties": {
        "issue_id": {
          "type": "string"
        },
        "closed": {
          "type": "integer"
        },
        "activity": {
          "type": "string"
        },
        "progress": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "platform": {
          "type": "integer"
        },
        "content": {
          "$ref": "#/definitions/Content"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      }
    },
    "Content": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "board_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "arena_id": {
          "type": "integer"
        },
        "comments": {
          "type": "integer"
        },
        "likers": {
          "type": "integer"
        },
        "scraps": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "imgs": {
          "type": "string"
        },
        "ip": {
          "type": "string"
        },
        "deleted": {
          "type": "integer"
        },
        "reported": {
          "type": "integer"
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        },
        "allow_comment": {
          "type": "integer"
        },
        "bodytype": {
          "type": "integer"
        },
        "celltype": {
          "type": "integer"
        },         
        "link": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      },
      "xml": {
        "name": "Content"
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserAction": {
      "type": "object",
      "properties": {
        "like": {
          "type": "integer"
        },
        "scrap": {
          "type": "integer"
        },
        "hasImgs": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "user_action"
      }
    },
...
    "ApiResponse2": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  },
...
}